+++ src/main/kotlin/com/ibrahimdans/i18n/extensions/lang/js/JsReferenceAssistant.kt
@@ -10,18 +10,14 @@ import com.ibrahimdans.i18n.plugin.key.FullKey
 import com.ibrahimdans.i18n.plugin.key.parser.KeyParserBuilder
 import com.ibrahimdans.i18n.plugin.utils.unQuote
 import com.intellij.lang.javascript.patterns.JSPatterns
-import com.intellij.lang.javascript.psi.JSCallExpression
 import com.intellij.lang.javascript.psi.ecma6.JSComputedPropertyNameOwner
 import com.intellij.lang.javascript.psi.JSConditionalExpression
 import com.intellij.lang.javascript.psi.JSLiteralExpression
 import com.intellij.lang.javascript.psi.JSProperty
-import com.intellij.lang.javascript.psi.JSReferenceExpression
-import com.intellij.lang.javascript.psi.JSThisExpression
 import com.intellij.lang.javascript.psi.ecma6.TypeScriptEnumField
 import com.intellij.patterns.ElementPattern
 import com.intellij.patterns.ElementPatternCondition
 import com.intellij.psi.PsiElement
-import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.psi.util.parents
 import com.intellij.util.ProcessingContext
 
@@ -55,13 +51,8 @@ internal class JsReferenceAssistant: ReferenceAssistant {
         }
     }
 
-    private fun isDirectOrConfiguredCall(element: PsiElement): Boolean {
-        val callExpr = PsiTreeUtil.getParentOfType(element, JSCallExpression::class.java) ?: return true
-        val refExpr = callExpr.methodExpression as? JSReferenceExpression ?: return true
-        val qualifier = refExpr.qualifier ?: return true
-        if (qualifier is JSThisExpression) return true
-        return refExpr.text in translationFunctionNames
-    }
+    private fun isDirectOrConfiguredCall(element: PsiElement): Boolean =
+        isDirectOrConfiguredCall(element, translationFunctionNames)
 
     override fun pattern(): ElementPattern<out PsiElement> =
         object : ElementPattern<PsiElement> {